refactor: colocate RouteGroupRow with its hook instead of a types file
Move the WIP RouteGroupRow type into src/api/route-groups.ts (the hook that
returns it) and delete the standalone src/types/route-group.ts. Consumers now
import RouteGroupRow from '@/api/route-groups'.

Why: when the backend grows the route-group aggregates (F4KRP-196), the
migration is then a one-file change — swap the local type for the generated
one and convert the hook — with no consumer churn and no file to delete,
unlike the address.ts/location.ts cleanup this PR already did. Also drops the
unused DeliveryType.
